refactors npexplorer api
