+++ config/apis.go
@@ -2,8 +2,8 @@ package config
 
 import _ "github.com/connectome-neuprint/neuPrintHTTP/api/dbmeta"
 import _ "github.com/connectome-neuprint/neuPrintHTTP/api/custom"
+import _ "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer"
 
-//import _ "github.com/connectome-neuprint/neuPrintHTTP/api/npexplorer"
 //import _ "github.com/connectome-neuprint/neuPrintHTTP/api/skeletons"
 //import _ "github.com/connectome-neuprint/neuPrintHTTP/api/raw/cypher"
 //import _ "github.com/connectome-neuprint/neuPrintHTTP/api/raw/spatial"
